Makefile, patches, and recompiled pdf.
It also fixes the line numbers that were appearing in the melba-accepted-special-issue.pdf

The idea is to apply dynamically, through the makefiles, the (small) patches for each custom pdf, and compile that one. This way, it is much easier to reflect in the three example pdfs changes in the main template.